重构(todo): 将QVariantMap替换为std::map<std::string, QVariant>并移除未使用的代码
移除未使用的QString和QDateTime头文件
将网络代理端口类型从int改为uint16_t
更新TODO.md文件，移除已完成的任务项
清理同步服务器中未使用的最后同步时间相关代码

+++ cpp/foundation/network_proxy.cpp
@@ -96,7 +96,7 @@ const QString &NetworkProxy::getProxyHost() const noexcept {
  * @brief 获取代理端口
  * @return 代理端口
  */
-constexpr int NetworkProxy::getProxyPort() const noexcept {
+constexpr uint16_t NetworkProxy::getProxyPort() const noexcept {
     return m_proxyPort;
 }
 
@@ -239,7 +239,7 @@ QNetworkProxy NetworkProxy::createQNetworkProxy() const noexcept {
             (m_proxyType == ProxyType::HttpProxy) ? QNetworkProxy::HttpProxy : QNetworkProxy::Socks5Proxy;
         proxy.setType(proxyType);
         proxy.setHostName(m_proxyHost);
-        proxy.setPort(static_cast<quint16>(m_proxyPort));
+        proxy.setPort((m_proxyPort));
 
         if (hasProxyAuth()) {
             proxy.setUser(m_proxyUsername);
@@ -260,7 +260,7 @@ QNetworkProxy NetworkProxy::createQNetworkProxy() const noexcept {
  * @param username 用户名
  * @param password 密码
  */
-void NetworkProxy::updateProxyConfig(ProxyType type, const QString &host, int port, const QString &username,
+void NetworkProxy::updateProxyConfig(ProxyType type, const QString &host, uint16_t port, const QString &username,
                                      const QString &password) noexcept {
     m_proxyType = type;
     m_proxyHost = host;
